Examples: Added example of extending the Python grammar
